Update navigation styling system
I decided to remove auxilary-nav block to improve resusability of nav
block. The difference between them was to small and could be covered
with just a new nav modifier instead of separate block.

---

Also for now I shifted the reset of nav--auxilary margins (the block is
used mainly with `ul` that has those by default, though it is not
prohibited to use any other HTML-element that may not have default
margins).

I've been thinking for a while whether blocks should reset defaults by
themselves or the ones should be overriten with mixes.

From the one point of view, it's quite convinient to remove initial
values of some properties that are defined by your browser in the block
class itself. Then if the block migrate it will not need to rely on
any global reset or it will not require to define new mixes that will
reset defaults.

But at least in the case of margins and positioning BEM recommend to use
mixes to define those types of properties. For now I will stick to BEM
recommendation and will search for a new information about the question.