Move nav-container to separate block
Off-canvas nav (type of nav currently implemented on sm screens) is a
standalone component. Previously it was relying too heavily on app
block.

This commit intended to separate nav-container in separate block that is
used with nav block (actually the former is the wrapper of the latter).

It's worth considering remove of app__container class that is used for
consistency of margins between UI elements and inside them.